% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamInformation.R
\name{addTeamLogo}
\alias{addTeamLogo}
\title{Loads svg images to each team}
\usage{
addTeamLogo(teamInfo)
}
\arguments{
\item{teamInfo}{Data frame of the team information}
}
\value{
Returns a data.frame with an image pointer of the svg logo
}
\description{
Loads svg images to each team
}
