% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{attKey}
\alias{attKey}
\title{Table of attribute names}
\format{
A data frame with 41 observations and 3 variables
\describe{
\item{abb}{Three letter abbreviation of the attribute names}
\item{attributeIndex}{Name of attribute from the Index (exported from FHM)}
\item{attribute}{Normal name of the attribute}
}
}
\usage{
attKey
}
\description{
A table containing the abbreviations, FHM, and normal attribute names.
}
\keyword{datasets}
