% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playerCard.R
\name{dataLoader}
\alias{dataLoader}
\title{Creates a function that loads the data used for the player cards}
\usage{
dataLoader(league, season = NULL)
}
\arguments{
\item{league}{League ID for the chosen league}

\item{season}{Season to take the data from}
}
\value{
Data frame of the chosen league and season
}
\description{
Creates a function that loads the data used for the player cards
}
