% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraperFunctions.R
\name{draftedProspectScraper}
\alias{draftedProspectScraper}
\title{Scrapes the prospects from each SHL team}
\usage{
draftedProspectScraper(link)
}
\arguments{
\item{link}{Forum link to the prospect pages for an SHL team}
}
\value{
Returns a data frame with each prospect player page link and the team holding their rights
}
\description{
Scrapes the prospects from each SHL team
}
