% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{forumData}
\alias{forumData}
\title{Forum Scrape Data}
\format{
A data frame with a number of observations and 41 variables
}
\usage{
forumData
}
\description{
An .RData containing the result of a daily forum scrape. It includes information of all players.
}
\keyword{datasets}
