% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shlAPI.R
\name{gamesLoader}
\alias{gamesLoader}
\title{Loads schedule from the Index}
\usage{
gamesLoader(leagueID, season = NULL, type = NULL)
}
\arguments{
\item{leagueID}{ID of the league to get attributes from, SHL = 0, SMJHL = 1, IIHF = 2, WJC = 3}

\item{season}{The season to get the attributes from}

\item{type}{Selects whether pre-season, regular season, or post-season}
}
\value{
Returns a data.frame with the scheduling
}
\description{
Loads schedule from the Index
}
\examples{
# Loads the SHL schedule from season 64
gamesLoader(leagueID = 0, season = 64, type = "regular")

}
