% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{historySkaterSeason}
\alias{historySkaterSeason}
\title{History Skater Data}
\format{
A data frame with a number of observations and variables
}
\usage{
historySkaterSeason
}
\description{
An .RData containing the historical data of all skaters from all four leagues (WJC is incomplete)
}
\keyword{datasets}
