% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shlAPI.R
\name{indStatsLoader}
\alias{indStatsLoader}
\title{Loading the player statistics}
\usage{
indStatsLoader(leagueID, season = NULL, type = NULL)
}
\arguments{
\item{leagueID}{ID of the league to get attributes from, SHL = 0, SMJHL = 1, IIHF = 2, WJC = 3}

\item{season}{The season to get the attributes from}

\item{type}{Selects whether pre-season, regular season, or post-season}
}
\value{
List of two data.frames, one consisting of players and one consisting of goalies
}
\description{
Loading the player statistics
}
\examples{
# Loads the SHL player statistics from S64
indStatsLoader(leagueID = 0, season = 64, type = "regular")
}
