% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playerCard.R
\name{playerCard}
\alias{playerCard}
\title{Creates a function that outputs a specific player card from a league}
\usage{
playerCard(chosen, leagueData)
}
\arguments{
\item{chosen}{Player chosen to visualize}

\item{leagueData}{The league data of the chosen league}
}
\value{
Returns an image of the player card
}
\description{
Creates a function that outputs a specific player card from a league
}
