% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playerCard.R
\name{playerData}
\alias{playerData}
\title{Creates a function that filters and processes the league data given a chosen player}
\usage{
playerData(chosen, leagueData)
}
\arguments{
\item{chosen}{Player chosen to visualize}

\item{leagueData}{The league data of the chosen league}
}
\value{
Returns a data frame with the data of the chosen player for the player card
}
\description{
Creates a function that filters and processes the league data given a chosen player
}
