% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shlAPI.R
\name{playerLoader}
\alias{playerLoader}
\title{Loading the player attributes}
\usage{
playerLoader(leagueID, season = NULL)
}
\arguments{
\item{leagueID}{ID of the league to get attributes from, SHL = 0, SMJHL = 1, IIHF = 2, WJC = 3}

\item{season}{The season to get the attributes from}
}
\value{
Returns a list of two data.frames, one for players and one for goalies
}
\description{
Loading the player attributes
}
\examples{
# Loads the SHL player attributes from S64
playerLoader(leagueID = 0, season = 64)
}
