% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shlAPI.R
\name{portalPlayers}
\alias{portalPlayers}
\title{Loads player data from the SHL Portal}
\usage{
portalPlayers()
}
\value{
Returns a data.frame with the information
}
\description{
Loads player data from the SHL Portal
}
