% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forumScraper.R
\name{prospectFALinks}
\alias{prospectFALinks}
\title{Combines all undrafted Prospects and Free Agents links}
\usage{
prospectFALinks()
}
\description{
Combines all undrafted Prospects and Free Agents links
}
