% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraperFunctions.R
\name{prospectsFAScraper}
\alias{prospectsFAScraper}
\title{Scrapes player links from undrafted prospecs or FA forum page}
\usage{
prospectsFAScraper(link)
}
\arguments{
\item{link}{The forum link}
}
\description{
Scrapes player links from undrafted prospecs or FA forum page
}
