% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forumScraper.R
\name{scraper}
\alias{scraper}
\title{The main forum scraper}
\usage{
scraper()
}
\value{
Returns a data frame with all player information from the forum.
}
\description{
Making use of created scraper functions, this function scrapes and aggregates all info from the forum
}
