% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shlAPI.R
\name{standingsLoader}
\alias{standingsLoader}
\title{Loads team standings from the Index}
\usage{
standingsLoader(leagueID, season = NULL)
}
\arguments{
\item{leagueID}{ID of the league to get attributes from, SHL = 0, SMJHL = 1, IIHF = 2, WJC = 3}

\item{season}{The season to get the attributes from}
}
\value{
Returns a data.frame with the information
}
\description{
Loads team standings from the Index
}
\examples{
# Loads the SHL team information from season 64
standingsLoader(leagueID = 0, season = 64)
}
