% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{teamInfo}
\alias{teamInfo}
\title{Table of all team information for the SHL leagues}
\format{
A data frame with 90+ observations and 12 variables
\describe{
\item{franchiseID}{ID value of the franchise}
\item{teamID}{ID value of the team, based on its creation in their respective league}
\item{fhmID}{The ID value that FHM uses for the team}
\item{team}{Team name}
\item{Inaugural.Season}{First season in their respective league}
\item{leagueID}{ID value of the league the team is playing in}
\item{league}{Name of the league the team is playing in}
\item{abbr}{Short abbreviation for the team}
\item{primary}{Primary color of the team}
\item{secondary}{Secondary color of the team}
\item{alt1}{Alternative color of the team}
\item{alt2}{Alternative color of the team}
}
}
\usage{
teamInfo
}
\description{
A data set containing information about the different teams in the 4 leagues of the SHL
}
\keyword{datasets}
