% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraperFunctions.R
\name{teamScraper}
\alias{teamScraper}
\title{Scrapes team links from a league}
\usage{
teamScraper(league)
}
\arguments{
\item{league}{The forum link to a league}
}
\description{
Scrapes team links from a league
}
