% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tpeCost}
\alias{tpeCost}
\title{Table of the TPE costs}
\format{
A data frame with 17 observations and 5 variables
\describe{
\item{Skill.level}{The value of the attribute, 4-20}
\item{TPE}{The TPE cost of having the attribute at that level, using the old scale}
\item{newTPE}{The TPE cost of having the attribute at that level, using the new scale}
\item{oldTPECost}{Cost of increasing value to this level, old scale}
\item{newTPECost}{Cost of increasing value to this level, new scale}
}
}
\usage{
tpeCost
}
\description{
A data set containing the costs for the different attributes values
using the old (pre-S61) TPE scale and the new.
}
\keyword{datasets}
