% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shlAPI.R
\name{tpeEarnings}
\alias{tpeEarnings}
\title{Reads and summarizes player earnings in a season from the SHL Portal}
\usage{
tpeEarnings(pid)
}
\value{
A data frame of earnings from the past seasons
}
\description{
Reads and summarizes player earnings in a season from the SHL Portal
}
