% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraperFunctions.R
\name{userScraper}
\alias{userScraper}
\title{Scrapes user data}
\usage{
userScraper(link)
}
\arguments{
\item{link}{Forum link to a user}
}
\value{
Data table with user data
}
\description{
Scrapes user data
}
