% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shlAPI.R
\name{readAPI}
\alias{readAPI}
\title{Calls the an SHL API}
\usage{
readAPI(url, ...)
}
\arguments{
\item{url}{The url to the SHL API either the Index or the Portal}

\item{...}{Additional queries for the API call.}
}
\value{
A data frame of the data.
}
\description{
Calls the an SHL API
}
\examples{
readAPI(
   url = "https://index.simulationhockey.com/api/v1/players/ratings",
   query = list(league = 0, season = 63)
)
}
